extract TimedConditionsCellRenderer into CommonEditor
